Remove inline.js
